+++ eras/shelley/test-suite/test/Test/Cardano/Ledger/Shelley/Examples/Mir.hs
@@ -37,12 +37,7 @@ import Cardano.Ledger.Shelley.LedgerState (
   emptyRewardUpdate,
  )
 import Cardano.Ledger.Shelley.Rules (
-  ShelleyBbodyPredFailure (..),
   ShelleyDelegPredFailure (..),
-  ShelleyDelegsPredFailure (..),
-  ShelleyDelplPredFailure (..),
-  ShelleyLedgerPredFailure (..),
-  ShelleyLedgersPredFailure (..),
   ShelleyUtxowPredFailure (..),
  )
 import Cardano.Ledger.Shelley.Tx (ShelleyTx (..))
@@ -243,15 +238,10 @@ mirFailWits pot =
     (initStMIR (Coin 1000))
     (blockEx1' insufficientMIRWits pot)
     ( Left
-        [ BbodyFailure @(ShelleyEra c)
-            ( LedgersFailure
-                ( LedgerFailure
-                    ( UtxowFailure $
-                        MIRInsufficientGenesisSigsUTXOW ws
-                    )
-                )
-            )
-        ]
+        . pure
+        . BbodyFailure
+        . injectFailure
+        $ MIRInsufficientGenesisSigsUTXOW ws
     )
   where
     ws = Set.fromList $ map (asWitness . aikColdKeyHash . coreNodeIssuerKeys) [0 .. 3]
@@ -271,22 +261,10 @@ mirFailFunds pot treasury llNeeded llReceived =
     (initStMIR treasury)
     (blockEx1' sufficientMIRWits pot)
     ( Left
-        [ BbodyFailure
-            ( LedgersFailure
-                ( LedgerFailure
-                    ( DelegsFailure
-                        ( DelplFailure
-                            ( DelegFailure $
-                                InsufficientForInstantaneousRewardsDELEG
-                                  pot
-                                  llNeeded
-                                  llReceived
-                            )
-                        )
-                    )
-                )
-            )
-        ]
+        . pure
+        . BbodyFailure
+        . injectFailure
+        $ InsufficientForInstantaneousRewardsDELEG pot llNeeded llReceived
     )
 
 --