+++ libs/small-steps-test/src/Control/State/Transition/Invalid/Trace.hs
@@ -12,6 +12,7 @@ module Control.State.Transition.Invalid.Trace where
 
 import Control.State.Transition (Environment, PredicateFailure, Signal, State)
 import qualified Control.State.Transition.Trace as Trace
+import Data.List.NonEmpty (NonEmpty)
 import GHC.Generics (Generic)
 import NoThunks.Class (NoThunks (..))
 
@@ -21,7 +22,7 @@ data Trace s = Trace
   -- ^ Last signal in the trace. This might cause a predicate failure, but it
   -- isn't guaranteed to do so, since invalid trace generation is
   -- probabilistic.
-  , errorOrLastState :: !(Either [PredicateFailure s] (State s))
+  , errorOrLastState :: !(Either (NonEmpty (PredicateFailure s)) (State s))
   }
   deriving (Generic)