+++ libs/small-steps-test/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Version history for `small-steps-test`
 
+## 1.1.0.0
+
+* Change `invalidSignalsAreGenerated`, `Trace`, `checkTrace`, `applySTSTest`, `traceFromInitState`,
+  `forAllTraceFromInitState`, `onlyValidSignalsAreGeneratedFromInitState`
+  to use `NonEmpty (PredicateFailure _)` instead of `[PredicateFailure _]`
+
 ## 1.0.1.1
 
 *